Allow builds on jenkins
Add some ci_tools to allow golden builds on jenkins.  Local
builds can still be done the same way as always, but making
golden builds involves upgrading the build version
(in <project>/BUILD) and pushing a commit to the repo.  This
will cause jenkins to build a new version of the package and
push to the repo.

In the near future, it will cause program tests to run first,
then push to the repo.

This change allows for both gated package pushes via jenkins,
as well as allowing in the future for multi-distro and
multi-arch builds